+++ template/src/plugins/cron_jobs/index.ts
@@ -1,5 +1,0 @@
-import { demoCronJob as demo } from './demo.job';
-
-export {
-    demo
-};